Splits model training into QoE and Tamper
